+++ pywatts/core/pipeline_step.py
@@ -41,10 +41,6 @@ class PipelineStep(Step):
             step.set_run_setting(run_setting)
         self.module.current_run_setting = self.current_run_setting
 
-    def _post_transform(self, result):
-        self.module._create_summary(self.current_run_setting.summary_formatter)
-        return super()._post_transform(result)
-
     def reset(self, keep_buffer=False):
         """
         Resets all information of the step concerning a specific run. Furthermore, it resets also all steps