+++ tests/unit/core/test_summary_step.py
@@ -30,7 +30,7 @@ class TestSummaryStep(unittest.TestCase):
         self.fm_mock = None
 
     def test_get_summary(self):
-        result = self.summary.get_summary()
+        result = self.summary.get_summary(None, None)
 
         self.step_mock.get_result.assert_called_once_with(None, None)
         self.module_mock.transform.assert_called_once_with(file_manager=self.fm_mock, input=self.result_mock)